# spring-ai-web/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Organisations","description":"Manage organisations in the system"},{"name":"ChatBots","description":"Manage Chatbots in the system"},{"name":"Conversations","description":"Manage conversations in the system"},{"name":"Training Data","description":"Manage training data in the system"}],"paths":{"/api/training-data":{"get":{"tags":["Training Data"],"operationId":"getTrainingData","parameters":[{"name":"request","in":"query","description":"Lazy loading request parameters","required":true,"schema":{"$ref":"#/components/schemas/LazyLoadingEventRequest"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageTrainingDataResource"}}}}}},"post":{"tags":["Training Data"],"operationId":"saveTrainingData","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainingDataResource"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainingDataResource"}}}}}}},"/api/organisations":{"post":{"tags":["Organisations"],"operationId":"saveOrganisation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrganizationResource"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrganizationResource"}}}}}}},"/api/conversations":{"get":{"tags":["Conversations"],"operationId":"getConversations","parameters":[{"name":"request","in":"query","description":"Lazy loading request parameters","required":true,"schema":{"$ref":"#/components/schemas/LazyLoadingEventRequest"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageConversationResource"}}}}}},"post":{"tags":["Conversations"],"operationId":"saveConversation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConversationResource"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConversationResource"}}}}}}},"/api/chatbots":{"get":{"tags":["ChatBots"],"operationId":"getAllChatbots","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChatbotResource"}}}}}}},"post":{"tags":["ChatBots"],"operationId":"saveChatbot","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatbotResource"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatbotResource"}}}}}}},"/api/chatbots/upload-document":{"post":{"tags":["ChatBots"],"operationId":"uploadDocument","parameters":[{"name":"botId","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK"}}}},"/api/chatbots/rag":{"post":{"tags":["ChatBots"],"operationId":"ragChat","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}}}}},"/api/training-data/{id}":{"get":{"tags":["Training Data"],"operationId":"getTrainingData_1","parameters":[{"name":"id","in":"path","description":"UUID of the training data","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainingDataResource"}}}}}},"delete":{"tags":["Training Data"],"operationId":"deleteTrainingData","parameters":[{"name":"id","in":"path","description":"UUID of the training data","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/training-data/search":{"get":{"tags":["Training Data"],"operationId":"searchTrainingData","parameters":[{"name":"query","in":"query","description":"Search query for training data autocomplete","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TrainingDataResource"}}}}}}}},"/api/organisations/{id}":{"get":{"tags":["Organisations"],"operationId":"getOrganisation","parameters":[{"name":"id","in":"path","description":"UUID of the organisation","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrganizationResource"}}}}}},"delete":{"tags":["Organisations"],"operationId":"deleteOrganisation","parameters":[{"name":"id","in":"path","description":"UUID of the organisation","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/organisations/search":{"get":{"tags":["Organisations"],"operationId":"searchOrganisations","parameters":[{"name":"query","in":"query","description":"Search query for organisation autocomplete","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrganizationResource"}}}}}}}},"/api/conversations/{id}":{"get":{"tags":["Conversations"],"operationId":"getConversation","parameters":[{"name":"id","in":"path","description":"UUID of the conversation","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConversationResource"}}}}}},"delete":{"tags":["Conversations"],"operationId":"deleteConversation","parameters":[{"name":"id","in":"path","description":"UUID of the conversation","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/conversations/search":{"get":{"tags":["Conversations"],"operationId":"searchConversations","parameters":[{"name":"query","in":"query","description":"Search query for conversation autocomplete","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ConversationResource"}}}}}}}},"/api/chatbots/{id}":{"get":{"tags":["ChatBots"],"operationId":"getChatbot","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatbotResource"}}}}}}},"/api/chatbots/paged":{"get":{"tags":["ChatBots"],"operationId":"findChatbots","parameters":[{"name":"request","in":"query","required":true,"schema":{"$ref":"#/components/schemas/LazyLoadingEventRequest"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageChatbotResource"}}}}}}},"/api/chatbots/{botId}":{"delete":{"tags":["ChatBots"],"operationId":"clearBotKnowledgeBase","parameters":[{"name":"botId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"ChatbotResource":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"description":{"type":"string"},"status":{"$ref":"#/components/schemas/ChatbotStatusEnum"},"organization":{"$ref":"#/components/schemas/OrganizationResource"},"systemPromptTemplate":{"type":"string"},"temperature":{"type":"number","format":"double"},"confidenceThreshold":{"type":"number","format":"double"},"responseTemplate":{"type":"string"},"metadataFilters":{"type":"object","additionalProperties":{"type":"string"}},"strictGuardrails":{"type":"boolean"},"customGuardrails":{"type":"string"}}},"ChatbotStatusEnum":{"type":"string","enum":["LIVE","DRAFT","ARCHIVED"]},"LanguageEnum":{"type":"string","enum":["fr","en","nl","de"]},"OrganizationResource":{"type":"object","description":"Organisation details","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"}}},"PriorityEnum":{"type":"string","enum":["TOP_PRIORITY","VERY_HIGH","HIGH","DEFAULT","LOW"]},"TrainingDataResource":{"type":"object","description":"Training data details","properties":{"id":{"type":"string","format":"uuid"},"title":{"type":"string"},"type":{"$ref":"#/components/schemas/TrainingDataTypeEnum"},"language":{"$ref":"#/components/schemas/LanguageEnum"},"priority":{"$ref":"#/components/schemas/PriorityEnum"},"contentDate":{"type":"string","format":"date-time"},"addedDate":{"type":"string","format":"date-time"},"chatbot":{"$ref":"#/components/schemas/ChatbotResource"}}},"TrainingDataTypeEnum":{"type":"string","enum":["WEBSITE","DOCUMENT","PDF","FAQ"]},"ConversationResource":{"type":"object","description":"Conversation details","properties":{"id":{"type":"string","format":"uuid"},"conversationId":{"type":"string"},"chatbot":{"$ref":"#/components/schemas/ChatbotResource"},"status":{"$ref":"#/components/schemas/ConversationStatusEnum"},"startTime":{"type":"string","format":"date-time"}}},"ConversationStatusEnum":{"type":"string","enum":["PENDING","REVIEWED"]},"ChatRequest":{"type":"object","properties":{"message":{"type":"string"},"conversationId":{"type":"string"},"botId":{"type":"string"}}},"ChatResponse":{"type":"object","properties":{"response":{"type":"string"},"sessionId":{"type":"string"},"confidence":{"type":"number","format":"double"},"retrievalConfidence":{"type":"number","format":"double"}}},"LazyLoadingEventRequest":{"type":"object","properties":{"first":{"type":"integer","format":"int32","description":"Index of the first record to fetch","example":0},"rows":{"type":"integer","format":"int32","description":"Number of records per page","example":10},"sortField":{"type":"string","description":"Field used for sorting","example":"name"},"sortOrder":{"type":"integer","format":"int32","description":"Sorting order, 1 for ascending, 0 for descending","example":1},"globalFilter":{"type":"string","description":"Global filter text to apply","example":"searchQuery"}},"required":["first","rows"]},"PageTrainingDataResource":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/TrainingDataResource"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"unpaged":{"type":"boolean"},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"}}},"SortObject":{"type":"object","properties":{"unsorted":{"type":"boolean"},"sorted":{"type":"boolean"},"empty":{"type":"boolean"}}},"PageConversationResource":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/ConversationResource"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"empty":{"type":"boolean"}}},"PageChatbotResource":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/ChatbotResource"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"empty":{"type":"boolean"}}}}}}