# spring-ai-web/src/app/@core/api/generated/models/priority-enum.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export enum PriorityEnum {
  TopPriority = 'TOP_PRIORITY',
  VeryHigh = 'VERY_HIGH',
  High = 'HIGH',
  Default = 'DEFAULT',
  Low = 'LOW'
}