# spring-ai-web/src/app/@core/api/generated/models/training-data-type-enum.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export enum TrainingDataTypeEnum {
  Website = 'WEBSITE',
  Document = 'DOCUMENT',
  Pdf = 'PDF',
  Faq = 'FAQ'
}